+++ android/build.gradle
@@ -35,5 +35,6 @@ android {
 
 dependencies {
   implementation 'com.android.billingclient:billing:1.1'
+  implementation files('jars/in-app-purchasing-2.0.76.jar')
 }
